Improved scanning for text boxes in DTPDF
I moved all the scanning of a content stream into CGPDFScanner. This collects the individual operations that it is listening for into DTPDFOperator instances. Then these are filtered into DTPDFTextBox instances where we try to get the proper positions of the boxes, though I don't know if the scaling is correct, so far I only use the tx and ty of the transforms set via TM.

The utility function to convert a CGPDF object into an NSObject is now together with some other C-functions in DTPDFFunctions. Note, I found with PDF strings that there are sometimes strings that cannot be easily decoded since they are comprised of lookup indexes, the used font determins this mode. Currently these are left as data and later on ignored