Make DTFoundation frameworks extension-safe
This replaces the extension-unsafe call to `[[UIApplication sharedApplication] statusBarOrientation]`
with a custom extension-safe extension on `UIScreen`. This in turn allows the dynamic frameworks
to be marked as using extension-safe APIs only which makes it possible to use them e.g. in an
iMessage app extension.

Based on https://stackoverflow.com/a/37329460